add adjacency utils
