+++ utils/__init__.py
@@ -4,6 +4,7 @@
 from .misc import *
 from .knn import *
 from .misc_cluster import *
+from .adjacency import *
 from .dataset import BasicDataset
 from .logger import create_logger
 from .faiss_search import faiss_search_knn
